requirements.txt
Flask==2.3.3
scikit-learn==1.5.1
spacy==3.6.0
rapidfuzz==3.1.1
requests==2.31.0
re2==1.0.0
ngrok==0.7.0